apps/api/requirements.txt
fastapi
uvicorn[standard]
pydantic
SQLAlchemy
psycopg2-binary
redis
python-dotenv
boto3
requests
